cleanup colors and change border color when tab will accept
