+++ src/vs/editor/contrib/inlineCompletions/browser/view/inlineEdits/gutterIndicatorMenu.ts
@@ -14,14 +14,15 @@ import { localize } from '../../../../../../nls.js';
 import { ICommandService } from '../../../../../../platform/commands/common/commands.js';
 import { IContextKeyService } from '../../../../../../platform/contextkey/common/contextkey.js';
 import { IKeybindingService } from '../../../../../../platform/keybinding/common/keybinding.js';
+import { asCssVariable, descriptionForeground, editorActionListForeground, editorHoverBorder } from '../../../../../../platform/theme/common/colorRegistry.js';
 import { Command } from '../../../../../common/languages.js';
 import { AcceptInlineCompletion, HideInlineCompletion, JumpToNextInlineEdit } from '../../controller/commands.js';
-import { ChildNode, FirstFnArg, LiveElement, n } from './utils.js';
+import { ChildNode, FirstFnArg, InlineEditTabAction, LiveElement, n } from './utils.js';
 
 export class GutterIndicatorMenuContent {
 	constructor(
 		private readonly _menuTitle: IObservable<string>,
-		private readonly _tabAction: IObservable<'jump' | 'accept' | 'inactive'>,
+		private readonly _tabAction: IObservable<InlineEditTabAction>,
 		private readonly _close: (focusEditor: boolean) => void,
 		private readonly _extensionCommands: IObservable<readonly Command[] | undefined>,
 		@IContextKeyService private readonly _contextKeyService: IContextKeyService,
@@ -54,7 +55,7 @@ export class GutterIndicatorMenuContent {
 		// TODO make this menu contributable!
 		return hoverContent([
 			header(this._menuTitle),
-			option(createOptionArgs({ id: 'gotoAndAccept', title: `${localize('goto', "Go To")} / ${localize('accept', "Accept")}`, icon: this._tabAction.map(action => action === 'accept' ? Codicon.check : Codicon.arrowRight), commandId: this._tabAction.map(action => action === 'accept' ? new AcceptInlineCompletion().id : new JumpToNextInlineEdit().id) })),
+			option(createOptionArgs({ id: 'gotoAndAccept', title: `${localize('goto', "Go To")} / ${localize('accept', "Accept")}`, icon: this._tabAction.map(action => action === InlineEditTabAction.Accept ? Codicon.check : Codicon.arrowRight), commandId: this._tabAction.map(action => action === 'accept' ? new AcceptInlineCompletion().id : new JumpToNextInlineEdit().id) })),
 			option(createOptionArgs({ id: 'reject', title: localize('reject', "Reject"), icon: Codicon.close, commandId: new HideInlineCompletion().id })),
 			separator(),
 			this._extensionCommands?.map(c => c && c.length > 0 ? [
@@ -87,7 +88,7 @@ function header(title: string | IObservable<string>) {
 	return n.div({
 		class: 'header',
 		style: {
-			color: 'var(--vscode-descriptionForeground)',
+			color: asCssVariable(descriptionForeground),
 			fontSize: '12px',
 			fontWeight: '600',
 			padding: '0 10px',
@@ -139,12 +140,12 @@ function separator() {
 	return n.div({
 		class: 'menu-separator',
 		style: {
-			color: 'var(--vscode-editorActionList-foreground)',
+			color: asCssVariable(editorActionListForeground),
 			padding: '2px 0',
 		}
 	}, n.div({
 		style: {
-			borderBottom: '1px solid var(--vscode-editorHoverWidget-border)',
+			borderBottom: `1px solid ${asCssVariable(editorHoverBorder)}`,
 		}
 	}));
 }
